Write testthat tests and examples together

Doctest lets you combine tests and examples. For more information, see @doctest and @expect.

Example

Example

Here's some documentation for a function:

#' Safe mean
#' 
#' @param x Numeric vector
#' @value The mean of `x`
#' 
#' @export
#'
#' @doctest
#' 
#' @expect equal(2)
#' safe_mean(1:3)
#' 
#' @expect warning("not numeric")
#' safe_mean("a")
#'
#' @expect warning("NA elements")
#' safe_mean(c(1, NA))
safe_mean <- function (x) {
  if (! is.numeric(x)) warning("x is not numeric")
  if (any(is.na(x))) warning("x contains NA elements")
  mean(x)
}

Instead of an @examples section, we have a @doctest section.

This will create tests like:

# File test-examples-safe_mean.R
# Generated by doctest: do not edit by hand

test_that("Example: safe_mean", {
  # Created from @examples for `safe_mean`
  # Source file: '<text>'
  # Source line: 9
  expect_equal(safe_mean(1:3), 2)
  expect_warning(safe_mean("a"), "not numeric")
  expect_warning(safe_mean(c(1, NA)), "NA elements")
})

The .Rd file will be created as normal, with an example section like:

\examples{
safe_mean(1:3)
safe_mean("a")
safe_mean(c(1, NA))
}

To use doctest in your package, add

Roxygen: list(roclets = c("collate", "rd", "namespace", "doctest::dt_roclet"))

to the DESCRIPTION file. You may also optionally add doctest to your 'Suggests:' dependencies.

Then run roxygen2::roxygenize() or devtools::document() from the command line.

Doctest is .

See also

Author

Maintainer: David Hugh-Jones davidhughjones@gmail.com